Atualizacao

+++ app.py
@@ -78,58 +78,74 @@ def pedindo_informacoes_para_o_usuario():
     session.commit()
 
 
-def exibindo_dados_usuario(cliente):
+def exibindo_dados_usuario():
     print("\nExibindo dados de todos os clientes.")
     lista_clientes = session.query(Cliente).all() #Consulta ao banco de dados "query" esconde o select - e vc seleciona a tabela que vc deseja que ele puxe
 
     for cliente in lista_clientes:
         print (f"{cliente.id} - {cliente.nome} - {cliente.email} - {cliente.senha} ")
 
+#Funcao para atualizar o cadastro, sobrepondo os dados
+def atualizar_cadastro(cliente):
+    nome_novo = input("Digite o novo nome: ")
+    email_novo = input("Digite o email novo: ")
+    senha_nova = input("Digite a senha nova: ")
+    if nome_novo:
+
+        cliente.nome = nome_novo
+    if email_novo:
+        cliente.email = email_novo
+    if senha_nova:
+        cliente.senha = senha_nova
+
+    session.commit()
+
+
 
 #Read - Select - Consulta
 pedindo_informacoes_para_o_usuario()
-exibindo_dados_usuario(Cliente)
+exibindo_dados_usuario()
 
 
 # U - Update - UPDATE - Atualizar
+os.system("cls || clear")
+escolha_atualizacao_cadastro = str(input("Deseja atualizar email do Cliente? "))
+if escolha_atualizacao_cadastro == "sim".lower(): 
+    email_cliente = input("Digite o email do cliente que será atualizado: ")
 
+    cliente = session.query(Cliente).filter_by(email=email_cliente).first()
 
-print("\nAtualizando dados do usuário.")
-email_cliente = input("Digite o email do cliente que será atualizado: ")
-
-cliente = session.query(Cliente).filter_by(email=email_cliente).first()
-
-if cliente:
-    pedindo_informacoes
-
-    session.commit()
-
+    if cliente:
+        atualizar_cadastro(cliente)
+    else:
+        print("Cliente não encontrado.")
 else:
-    print("Cliente não encontrado.")
-
+    pass
 # R - Read  - SELECT - Consulta
 
 print("\nExibindo dados de todos os clientes.")
-lista_clientes = session.query(Cliente).all() #Consulta ao banco de dados "query" esconde o select - e vc seleciona a tabela que vc deseja que ele puxe
-
-for cliente in lista_clientes:
-    exibindo_dados
+exibindo_dados_usuario()
 
 
 # D - Delete - DELETE - Excluir
+os.system("cls || clear")
+escolha = str(input("Deseja excluir um cliente? Sim ou Nao: "))
+if escolha == "sim".lower() :
 
-print("\nExcluindo os dados de um cliente.")
-email_cliente = input("Digite o e-mail do cliente que será excluído: ")
+    print("\nExcluindo os dados de um cliente.")
+    email_cliente = input("Digite o e-mail do cliente que será excluído: ")
 
-cliente = session.query(Cliente).filter_by(email = email_cliente).first()
+    cliente = session.query(Cliente).filter_by(email = email_cliente).first()
 
-if cliente:
-    session.delete(cliente)
-    session.commit()
-    print(f"Cliente {cliente.nome} excluído com sucesso!")
+    if cliente:
+        session.delete(cliente)
+        session.commit()
+        print(f"Cliente {cliente.nome} excluído com sucesso!")
 
+    else:
+        print("Cliente não encontrado.")
 else:
-    print("Cliente não encontrado.")
+    pass
 
 
 #R - Read - SELECT - Consulta
@@ -138,24 +154,29 @@ print("\nExibindo dados de todos os clientes.")
 lista_clientes = session.query(Cliente).all() #Consulta ao banco de dados "query" esconde o select - e vc seleciona a tabela que vc deseja que ele puxe
 
 for cliente in lista_clientes:
-    exibindo_dados
+    exibindo_dados_usuario
 
-
-
+os.system("cls || clear")
+consulta_cadastro_cliente = str(input("Deseja realizar consulta? "))
 # R - Read - SELECT - Consulta
+if consulta_cadastro_cliente == "sim".lower():
 
-print("Consultando os dados de apenas um cliente.")
-email_cliente = input("Digite o e-mail do cliente")
+    print("Consultando os dados de apenas um cliente.")
+    email_cliente = input("Digite o e-mail do cliente: ")
 
-cliente = session.query(Cliente).filter_by(email = email_cliente).first()
+    cliente = session.query(Cliente).filter_by(email = email_cliente).first()
 
-if cliente:
-     exibindo_dados
+    if cliente:
+        exibindo_dados_usuario()
 
+    else:
+        print("Cliente não encontrado.")
 else:
-    print("Cliente não encontrado.")
-
+    pass
 
 #Fechando conexão
+os.system("cls || clear")
+exibindo_dados_usuario()
+
 
 session.close()
